+++ piecrust/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change `Session::feeder_call` and `Session::feeder_call_raw` to allow for the
+  caller to specify the gas limit [#344]
+
 ## [0.17.0] - 2024-02-28
 
 ### Added
@@ -379,6 +384,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ISSUES -->
 
+[#344]: https://github.com/dusk-network/piecrust/issues/344
 [#336]: https://github.com/dusk-network/piecrust/issues/336
 [#325]: https://github.com/dusk-network/piecrust/issues/325
 [#324]: https://github.com/dusk-network/piecrust/issues/324